+++ src/GitHub_Updater/Base.php
@@ -487,6 +487,7 @@ class Base {
 		$this->tag     = isset( $_GET['rollback'] ) ? sanitize_text_field( wp_unslash( $_GET['rollback'] ) ) : false;
 		$slug          = 'plugin' === $type ? $repo->file : $repo->slug;
 		$download_link = $repo_api->construct_download_link( $this->tag );
+		add_filter( 'http_request_args', [ $this, 'download_package' ], 10, 2 );
 
 		/**
 		 * Filter download link so developers can point to specific ZipFile