+++ CHANGES.md
@@ -3,6 +3,8 @@
 * switching away from `master` or tag will use that branch for updating, not the release asset
 * update to JS to work with IE11, thanks @sharevb, arrow functions not supported in IE11
 * removed filter `github_updater_hide_branch_switcher` in favor of better branch switching
+* direct injection of authentication headers into `wp_remote_get()`
+* filter added for adding authentication headers for downloads packages
 
 #### 9.4.1 / 2020-04-04 **HotFix**
 * fixed problem with sanitizing